requirements.txt
# Medical-SchemaKG Framework - Python Dependencies

# Core dependencies
networkx>=3.0
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.28.0

# Optional dependencies for visualization
matplotlib>=3.5.0
rdflib>=6.0.0

# Utility dependencies
typing-extensions>=4.0.0
